% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxBlur.r
\name{BoxBlur}
\alias{BoxBlur}
\title{BoxBlur}
\usage{
BoxBlur(mat.mtx, box.num = NULL, boxSize.num = NULL, sd.num = 1)
}
\arguments{
\item{mat.mtx}{<matrix>: Numerical matrix}

\item{box.num}{<numeric>: the numerical vector for kernel. If Null apply a GaussBox (see 'GaussBox' function)}

\item{boxSize.num}{<numeric>: If box.num is NULL, size of kernel for 'GaussBox' function.}

\item{sd.num}{<numeric>: If box.num is NULL,  standard deviation parameter for 'GaussBox' function.}
}
\value{
Blurred matrix
}
\description{
Apply a blur on matrix with a One dimensional kernel
}
\examples{
set.seed(981643)
mat.mtx <- rnorm(10000,50,10)**3 \%>\% matrix(.,100,100)
pdf(file=paste0(getwd(),"/Rplots.pdf"))
heatmap(mat.mtx,Rowv=NA,Colv=NA)
heatmap(BoxBlur(mat.mtx),Rowv=NA,Colv=NA)
dev.off()
}
