% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakVector.r
\name{BreakVector}
\alias{BreakVector}
\title{BreakVector}
\usage{
BreakVector(
  x.num = NULL,
  min.num = NULL,
  center.num = NULL,
  max.num = NULL,
  n.num = 10,
  method.chr = "linear"
)
}
\arguments{
\item{x.num}{<numeric>: numerical vector}

\item{min.num}{<numeric>: minimal fixed value}

\item{center.num}{<numeric>: center fixed value}

\item{max.num}{<numeric>: maximal fixed value}

\item{n.num}{<numeric>: number of tile (n.num+1 breaks return)}

\item{method.chr}{<character>: kind of breaking. "linear" or "density". (Default "linear")}
}
\value{
Numerical vector of breaks
}
\description{
Compute the n+1 breaks of a vector in a linear or density based way with the possibility to fix minimal, center and maxiaml values.
}
\examples{
set.seed(31415)
x.num <- rnorm(100,50,200)
BreakVector(x.num=x.num, n.num=9)
}
