% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CbrtNorm.r
\name{CbrtNorm}
\alias{CbrtNorm}
\title{CbrtNorm}
\usage{
CbrtNorm(x.num, a.num = 0)
}
\arguments{
\item{x.num}{<numeric>: Numerical vector}

\item{a.num}{<numeric>: Constant value to add before compute cubic root (Default 0)}
}
\value{
Normalized vector
}
\description{
Normalize data by cubic root and adding a constant value
}
\examples{
set.seed(542972)
x.num <- rnbinom(1000, 1, 0.01)
pdf(file=paste0(getwd(),"/Rplots.pdf"))
    plot(density(x.num))
    plot(density(CbrtNorm(x.num, abs(min(x.num)))))
dev.off()
}
