% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log2Norm.r
\name{Log2Norm}
\alias{Log2Norm}
\title{Log2Norm}
\usage{
Log2Norm(x.num, a.num = 0)
}
\arguments{
\item{x.num}{<numeric>: Numerical vector}

\item{a.num}{<numeric>: Constant value to add before compute cubic root (Default 0)}
}
\value{
Normalized vector
}
\description{
Normalize data by logarithme (base 2) and adding a constant value
}
\examples{
set.seed(542972)
x.num <- rnbinom(1000, 1, 0.01)
pdf(file=paste0(getwd(),"/Rplots.pdf"))
    plot(density(x.num))
    plot(density(Log2Norm(x.num, abs(min(x.num)))))
dev.off()
}
