% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianSkewness.r
\name{MedianSkewness}
\alias{MedianSkewness}
\title{MedianSkewness}
\usage{
MedianSkewness(x.num)
}
\arguments{
\item{x.num}{<numeric>: Numerical vector}
}
\value{
Skewness value
}
\description{
Pearson Median Skewness
}
\examples{
set.seed(542972)
MedianSkewness(rnbinom(10000, 1, 0.01))
set.seed(542972)
MedianSkewness(rnorm(10000, 1, 0.01))
}
