% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.r
\name{Normalize}
\alias{Normalize}
\title{Normalize}
\usage{
Normalize(x.num = NULL, a.num = NULL)
}
\arguments{
\item{x.num}{<numeric>: Numerical vector}

\item{a.num}{<numeric>: Constant value to add before compute normalisation too avoid negativ number. If null auto dertermine a.num. (Default NULL)}
}
\value{
Normalized vector
}
\description{
Normalize data by with square root, cubic root or log2 based on minimal skewness
}
\examples{
set.seed(542972)
x.num <- rnbinom(1000, 1, 0.01)
pdf(file=paste0(getwd(),"/Rplots.pdf"))
    plot(density(x.num))
    plot(density(Normalize(x.num)))
dev.off()
}
